requirements.txt
snowflake-connector-python>=3.0.0
pandas>=2.0.0
numpy>=1.24.0
xgboost>=2.0.0
scikit-learn>=1.3.0
streamlit>=1.28.0
plotly>=5.18.0
python-dotenv>=1.0.0
